Add docs input `Color::class`.
